+++ iroh-sync/src/store/memory.rs
@@ -8,6 +8,7 @@ use std::{
 
 use anyhow::Result;
 use derive_more::From;
+use ed25519_dalek::{SignatureError, VerifyingKey};
 use parking_lot::{RwLock, RwLockReadGuard};
 
 use crate::{
@@ -304,18 +305,8 @@ pub struct ReplicaStoreInstance {
 }
 
 impl PublicKeyStore for ReplicaStoreInstance {
-    fn namespace_key(
-        &self,
-        bytes: &NamespaceId,
-    ) -> std::result::Result<NamespacePublicKey, ed25519_dalek::SignatureError> {
-        self.store.pubkeys.namespace_key(bytes)
-    }
-
-    fn author_key(
-        &self,
-        bytes: &AuthorId,
-    ) -> std::result::Result<AuthorPublicKey, ed25519_dalek::SignatureError> {
-        self.store.pubkeys.author_key(bytes)
+    fn public_key(&self, id: &[u8; 32]) -> std::result::Result<VerifyingKey, SignatureError> {
+        self.store.pubkeys.public_key(id)
     }
 }
